apu forms validated, apu activities done

+++ backend/database/migrations/2023_07_08_164231_create_apu_activites_table.php
@@ -15,7 +15,7 @@ class CreateApuActivitesTable extends Migration
     {
         Schema::create('apu_activities', function (Blueprint $table) {
             $table->id();
-            $table->integer('cap');
+            $table->double('cap');
             $table->text('description');
             $table->string('unit');
             $table->integer('quantity');
